refactor: common 패키지 구조 수정 - common.{계층}.{도메인} -> common.{도메인}.{계층}
