+++ src/main/java/kr/hhplus/be/server/common/queue/application/QueueTokenInfo.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.application.queue;
+package kr.hhplus.be.server.common.queue.application;
 
 public record QueueTokenInfo(String token, Long rank, Long totalWaitMinute, QueueTokenStatus status) {
