+++ src/main/java/kr/hhplus/be/server/common/queue/application/QueueTokenStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.application.queue;
+package kr.hhplus.be.server.common.queue.application;
 
 /**
  * 대기열 토큰 상태 Enum
