+++ src/test/java/kr/hhplus/be/server/common/lock/infrastructure/ReentrantLockManagerTest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.infrastructure.lock;
+package kr.hhplus.be.server.common.lock.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,7 +10,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import kr.hhplus.be.server.common.application.lock.LockManager;
+import kr.hhplus.be.server.common.lock.application.LockManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
