+++ src/test/java/kr/hhplus/be/server/common/queue/infrastructure/RedisQueueTokenManagerTest.java
@@ -1,10 +1,10 @@
-package kr.hhplus.be.server.common.infrastructure.queue;
+package kr.hhplus.be.server.common.queue.infrastructure;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 import java.util.Objects;
-import kr.hhplus.be.server.common.application.queue.QueueTokenInfo;
-import kr.hhplus.be.server.common.application.queue.QueueTokenStatus;
+import kr.hhplus.be.server.common.queue.application.QueueTokenInfo;
+import kr.hhplus.be.server.common.queue.application.QueueTokenStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
